backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.67.0
anyio==3.7.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
distro==1.9.0
docstring-parser==0.15
fastapi==0.104.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
instructor==0.4.8
jiter==0.10.0
livekit==1.0.13
livekit-api==1.0.5
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.6.4
numpy==2.3.3
openai==1.107.2
packaging==25.0
propcache==0.3.2
protobuf==6.32.1
pydantic==2.5.0
python-dotenv==1.1.1
pydantic_core==2.14.1
Pygments==2.19.2
PyYAML==6.0.2
requests==2.32.5
rich==13.9.4
sniffio==1.3.1
starlette==0.27.0
tokenizers==0.22.0
tqdm==4.67.1
typer==0.9.4
types-protobuf==6.30.2.20250822
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.24.0
vosk
websockets
httpx
yarl==1.20.1
